pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a11y-remediate"
version = "0.1.0"
description = "AI-powered WCAG 2.1 AA accessibility remediation for university course materials"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "lxml>=5.0",
    "Pillow>=10.0",
    "wcag-contrast-ratio>=0.9",
    "pydantic>=2.0",
    "google-genai>=1.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.7",
    "pymupdf>=1.24",
    "weasyprint>=68.0",
    "bcrypt>=4.0",
    "pyjwt>=2.8",
    "authlib>=1.3",
    "httpx>=0.27",
]

[project.optional-dependencies]
validation = [
    "playwright>=1.40",
    "axe-playwright-python>=0.1.3",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
